azure-iceberg-sink/requirements.txt
quixstreams==2.11.1
python-dotenv
pyarrow
pyiceberg
boto3
mypy_boto3_glue
sqlalchemy
psycopg2-binary
adlfs
numpy
redis